requirements.txt
torch==2.3.1
pytorch-lightning==2.0.2
torchmetrics==1.4.0
fair-esm==2.0.0
scikit-learn
numpy==1.26.4
scipy==1.11.3
pandas==2.2.2
biopython==1.84
wandb==0.18.2